refactor: Move non-public utilities to their own file
