+++ docs/README.md
@@ -7,5 +7,7 @@ generate-runtypes
 ### Modules
 
 - [\_\_tests\_\_/main.test](modules/__tests___main_test.md)
+- [\_\_tests\_\_/util.test](modules/__tests___util_test.md)
 - [main](modules/main.md)
 - [types](modules/types.md)
+- [util](modules/util.md)
