+++ docs/modules/types.md
@@ -28,7 +28,7 @@
 
 Ƭ **AnyType**: *rt.Static*<*typeof* anyTypeRt\>
 
-Defined in: [src/types.ts:162](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L162)
+Defined in: [src/types.ts:162](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L162)
 
 ___
 
@@ -44,7 +44,7 @@ ___
 | `readonly?` | *boolean* |
 | `type` | [*AnyType*](types.md#anytype) |
 
-Defined in: [src/types.ts:78](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L78)
+Defined in: [src/types.ts:78](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L78)
 
 ___
 
@@ -59,7 +59,7 @@ ___
 | `kind` | ``"dictionary"`` |
 | `valueType` | [*AnyType*](types.md#anytype) |
 
-Defined in: [src/types.ts:95](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L95)
+Defined in: [src/types.ts:95](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L95)
 
 ___
 
@@ -74,7 +74,7 @@ ___
 | `kind` | ``"intersect"`` |
 | `types` | [*AnyType*](types.md#anytype)[] |
 
-Defined in: [src/types.ts:130](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L130)
+Defined in: [src/types.ts:130](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L130)
 
 ___
 
@@ -82,7 +82,7 @@ ___
 
 Ƭ **LiteralType**: *rt.Static*<*typeof* literalTypeRt\>
 
-Defined in: [src/types.ts:32](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L32)
+Defined in: [src/types.ts:32](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L32)
 
 ___
 
@@ -90,7 +90,7 @@ ___
 
 Ƭ **NamedType**: *rt.Static*<*typeof* namedTypeRt\>
 
-Defined in: [src/types.ts:46](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L46)
+Defined in: [src/types.ts:46](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L46)
 
 ___
 
@@ -108,7 +108,7 @@ ___
 | `readonly?` | *boolean* |
 | `type` | [*AnyType*](types.md#anytype) |
 
-Defined in: [src/types.ts:48](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L48)
+Defined in: [src/types.ts:48](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L48)
 
 ___
 
@@ -123,7 +123,7 @@ ___
 | `fields` | [*RecordField*](types.md#recordfield)[] |
 | `kind` | ``"record"`` |
 
-Defined in: [src/types.ts:56](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L56)
+Defined in: [src/types.ts:56](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L56)
 
 ___
 
@@ -131,7 +131,7 @@ ___
 
 Ƭ **RootType**: *rt.Static*<*typeof* [*rootTypeRt*](types.md#roottypert)\>
 
-Defined in: [src/types.ts:174](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L174)
+Defined in: [src/types.ts:174](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L174)
 
 ___
 
@@ -139,7 +139,7 @@ ___
 
 Ƭ **SimpleType**: *rt.Static*<*typeof* simpleTypeRt\>
 
-Defined in: [src/types.ts:21](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L21)
+Defined in: [src/types.ts:21](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L21)
 
 ___
 
@@ -154,7 +154,7 @@ ___
 | `kind` | ``"union"`` |
 | `types` | [*AnyType*](types.md#anytype)[] |
 
-Defined in: [src/types.ts:110](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L110)
+Defined in: [src/types.ts:110](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L110)
 
 ## Variables
 
@@ -162,4 +162,4 @@ Defined in: [src/types.ts:110](https://github.com/cobraz/generate-runtypes/blob/
 
 • `Const` **rootTypeRt**: *Intersect*<[*Record*<{ `name`: *String* ; `type`: *Union*<[*Runtype*<[*ArrayType*](types.md#arraytype)\>, *Runtype*<[*DictionaryType*](types.md#dictionarytype)\>, *Runtype*<[*IntersectionType*](types.md#intersectiontype)\>, *Record*<{ `kind`: *Literal*<``"literal"``\> ; `value`: *Union*<[*Boolean*, *Literal*<``null``\>, *Number*, *String*, *Literal*<undefined\>]\>  }, ``false``\>, *Record*<{ `kind`: *Literal*<``"named"``\> ; `name`: *String*  }, ``false``\>, *Runtype*<[*RecordType*](types.md#recordtype)\>, *Record*<{ `kind`: *Union*<[*Literal*<``"boolean"``\>, *Literal*<``"function"``\>, *Literal*<``"never"``\>, *Literal*<``"null"``\>, *Literal*<``"number"``\>, *Literal*<``"string"``\>, *Literal*<``"symbol"``\>, *Literal*<``"undefined"``\>, *Literal*<``"unknown"``\>]\>  }, ``false``\>, *Runtype*<[*UnionType*](types.md#uniontype)\>]\>  }, ``false``\>, *InternalRecord*<{ `comment`: *Union*<[*String*, *Arr*<String, ``false``\>]\> ; `export`: *Boolean*  }, ``true``, ``false``\>]\>
 
-Defined in: [src/types.ts:164](https://github.com/cobraz/generate-runtypes/blob/5e188fd/src/types.ts#L164)
+Defined in: [src/types.ts:164](https://github.com/cobraz/generate-runtypes/blob/1c53623/src/types.ts#L164)
