+++ src/__tests__/main.test.ts
@@ -1,4 +1,4 @@
-import { generateRuntypes, groupFieldKinds } from '../main';
+import { generateRuntypes } from '../main';
 import { RootType } from '../types';
 
 describe('runtype generation', () => {
@@ -148,26 +148,6 @@ describe('runtype generation', () => {
   });
 
   describe('objects', () => {
-    it('groupFieldKinds', () => {
-      const res = groupFieldKinds([
-        { name: 'field_1', type: { kind: 'string' } },
-        { name: 'field_2', type: { kind: 'string' }, nullable: true },
-        { name: 'field_3', type: { kind: 'string' }, readonly: true },
-        {
-          name: 'field_4',
-          type: { kind: 'string' },
-          nullable: true,
-          readonly: true,
-        },
-      ]);
-
-      const [default_, nullable, readonly, both] = res;
-      expect(default_.fields.map((e) => e.name)).toEqual(['field_1']);
-      expect(nullable.fields.map((e) => e.name)).toEqual(['field_2']);
-      expect(readonly.fields.map((e) => e.name)).toEqual(['field_3']);
-      expect(both.fields.map((e) => e.name)).toEqual(['field_4']);
-    });
-
     it('default modifiers', async () => {
       const raw = generateRuntypes({
         name: 'test',
