+++ src/main.ts
@@ -11,6 +11,7 @@ import {
   UnionType,
   rootTypeRt,
 } from './types';
+import { groupFieldKinds } from './util';
 
 export type {
   PrettierOptions,
@@ -284,41 +285,6 @@ function writeComment(w: CodeWriter, comment: string | string[]) {
     }
     w.write(`*/\n`);
   }
-}
-
-/**
- * public for testing
- *
- * Used to evaluate if `Record` type include `readonly` and/or `nullable`
- * @private
- */
-export function groupFieldKinds(fields: readonly RecordField[]): {
-  readonly: boolean;
-  nullable: boolean;
-  fields: RecordField[];
-}[] {
-  return [
-    {
-      readonly: false,
-      nullable: false,
-      fields: fields.filter((e) => !e.readonly && !e.nullable),
-    },
-    {
-      readonly: false,
-      nullable: true,
-      fields: fields.filter((e) => !e.readonly && e.nullable),
-    },
-    {
-      readonly: true,
-      nullable: false,
-      fields: fields.filter((e) => e.readonly && !e.nullable),
-    },
-    {
-      readonly: true,
-      nullable: true,
-      fields: fields.filter((e) => e.readonly && e.nullable),
-    },
-  ].filter((e) => e.fields.length > 0);
 }
 
 function writeRecordType(
