+++ src/util.ts
@@ -1,0 +1,36 @@
+import { RecordField } from './types';
+
+/**
+ * public for testing
+ *
+ * Used to evaluate if `Record` type include `readonly` and/or `nullable`
+ * @private
+ */
+export function groupFieldKinds(fields: readonly RecordField[]): {
+  readonly: boolean;
+  nullable: boolean;
+  fields: RecordField[];
+}[] {
+  return [
+    {
+      readonly: false,
+      nullable: false,
+      fields: fields.filter((e) => !e.readonly && !e.nullable),
+    },
+    {
+      readonly: false,
+      nullable: true,
+      fields: fields.filter((e) => !e.readonly && e.nullable),
+    },
+    {
+      readonly: true,
+      nullable: false,
+      fields: fields.filter((e) => e.readonly && !e.nullable),
+    },
+    {
+      readonly: true,
+      nullable: true,
+      fields: fields.filter((e) => e.readonly && e.nullable),
+    },
+  ].filter((e) => e.fields.length > 0);
+}
